DoubleOracle

Documentation for DoubleOracle.

DoubleOracle.ActionSet

DoubleOracle.ActionSet

DoubleOracle.Algorithm

DoubleOracle.Game

DoubleOracle.LinearProgram

DoubleOracle.MatrixGame

DoubleOracle.MatrixGame

DoubleOracle.NormalFormGame

DoubleOracle.Player

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.iterate

Base.length

DoubleOracle.allnames

DoubleOracle.createplayers

DoubleOracle.load

DoubleOracle.solve

Documentation

DoubleOracle.ActionSet — Type

ActionSet

Set of actions of a single player. Store names and assign ids for internal representation.

Fields

player: Player owning the set of actions

   n: Number of actions in the set

 ids: Vector of assigned integer ids

idtoname: Map translating an integer id to an action name

nametoid: Map translating name of action to the assigned id

DoubleOracle.ActionSet — Method

ActionSet(player, actionnames)

Construct ActionSet for player containing action with names as in actionnames.

Examples

julia> ActionSet(Player(1), ["A", "B", "C"])
Actions of Player 1: [1] A | [2] B | [3] C |

julia> ActionSet(Player(2), ["f", "e", "d"])
Actions of Player 2: [1] f | [2] e | [3] d |

DoubleOracle.Algorithm — Type

Algorithm

Abstract type representing solving algorithms of a game in the Game Theory

DoubleOracle.Game — Type

Game

Abstract type uniting all types of games in Game Theory.

DoubleOracle.LinearProgram — Type

LinearProgram

Subtype of Algorithm representing exact solution of a Normal-Form game by linear programming.

DoubleOracle.MatrixGame — Type

MatrixGame

Solve and represent solution of a Matrix game u. Solution is made by a standard linear program for two-player zero-sum Normal-form games. Both outcome and an equilibrium strategy is saved for both players.

Fields

outcomes: 2-tuple with game outcomes for each respective player

strategies: 2-tuple with en equilibrium strategies (probability distribution over rows/columns Δ(A)) for each respective player

Example

julia> u = [1 -1; -1 1];
julia> MatrixGame(u)
MatrixGame results:
→ outcome of the Nash Equilibrium: (0.0, -0.0)
→ strategy of row player: [0.5, 0.5]
→ strategy of columne player: [0.5, 0.5]

DoubleOracle.MatrixGame — Method

(MatrixGame)(player)

Obtain the outcome value of a player from the MatrixGame structure.

Example

julia> mg = MatrixGame([1 -1; -1 1]);
julia> mg(Player(1))
0.0

DoubleOracle.NormalFormGame — Type

NormalFormGame{T}

Model for a two-player zero-sum game in Normal Form. Here, player 1 tries to maximise the outcome of type T, player 2 minimises it.

Fields

name: Name of the NF game

N: Players of the game

A: Available actions of each player

U: Game matrix containing the outcomes for each joint action profile

DoubleOracle.Player — Type

Player

Represent a Player with an integer id in a Game

Examples

julia> Player(1)
Player 1

julia> Player(2)
Player 2

Base.getindex — Method

getindex(A, id::Integer)

Retrieve name of action with assigned id.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> A[3]
"C"

Base.getindex — Method

getindex(A, name::String)

Retrieve id of action named name.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> A["B"]
2

Base.getindex — Method

getindex(MatrixGame, player)

Obtain the strategy of a player from the MatrixGame structure.

Example

julia> mg = MatrixGame([1 -1; -1 1]);
julia> mg[Player(1)]
2-element Vector{Float64}:
 0.5
 0.5

Base.getindex — Method

getindex(nfg, a1, a2)

Obtain nfg outcome by playing joing action profile (a1, a2)

Example

julia> nfg = load("./data/nf_games/mathing_pennies.nfg, NormalFormGame);
julia> nfg[1, 1]
1

julia> nfg["1", "B"]
-1

Base.getindex — Method

getindex(container, player)

Get the element on index of a player inside the container. The container has to be either Vector or Tuple The index corresponds to id of the player.

Examples

julia> names = ("John", "Thomas", "Agatha")
("John", "Thomas", "Agatha")
julia> names[Player(3)]
"Agatha"

Base.iterate — Function

iterate(A)

Go through ids of all actions in fixed order.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> collect(A)
4-element Vector{Any}:
 1
 2
 3
 4

Base.length — Method

length(A)

Return number of actions in A.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> length(A)
4

DoubleOracle.allnames — Method

allnames(A)

Return list of action names from A in the same order as ids

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> allnames(A)
["A", "B", "C", "D"]

DoubleOracle.createplayers — Method

createplayers(n)

Create a vector of $n$ Players with ids $i ∈ {1,…,n}$

Examples

julia> createplayers(3)
3-element Vector{Player}:
 Player 1
 Player 2
 Player 3

DoubleOracle.load — Method

load(filepath, type)

Load and return a game of type (i.e. $NormalFormGame$, ...) from file located in the filesystem at filepath. The extension of the filepath must correspond to the game type (e.g. $NormalFormGame -> .nfg$).

EXample

julia> nfg = load("./data/nf_games/mathing_pennies.nfg, NormalFormGame)

DoubleOracle.solve — Method

solve(NormalFormGame, LinearProgram)

Solve the NormalFormGame by linear programming and return the outcome and equilibrium strategies.