Game

This parts documents loading and generating a new game and auxiliary functions for this process.

Player

Structure and functions representing Players in the game. Player is used mainly for printing and for indexing in more complicated structures.

One or more Players can be created at once.

DoubleOracle.Player — Type

Player

Represent a Player with an integer id in a Game

Examples

julia> Player(1)
Player 1

julia> Player(2)
Player 2

DoubleOracle.createplayers — Method

createplayers(n::Integer)

Create a vector of $n$ Players with ids $i ∈ {1,…,n}$

Examples

julia> createplayers(3)
3-element Vector{Player}:
 Player 1
 Player 2
 Player 3

As said above, a Player structure can be used to easier indexing without the need of remembering ids.

Base.getindex — Method

getindex(container::AbstractVecOrTuple, player::Player)

Get the element on index of a player inside the container. The container has to be either Vector or Tuple The index corresponds to id of the player.

Examples

julia> names = ("John", "Thomas", "Agatha")
("John", "Thomas", "Agatha")

julia> names[Player(3)]
"Agatha"

ActionSet

Structure that provides an abstraction over a set of available actions of a single player. It gives easy translations between names given by user in the input file and ids that are used as indices to strategy vectors and utility matrix.

DoubleOracle.ActionSet — Type

ActionSet

Set of actions of a single player. Store names and assign ids for internal representation.

Fields

player: Player owning the set of actions

n: Number of actions in the set

ids: Vector of assigned integer ids

idtoname: Map translating an integer id to an action name

nametoid: Map translating name of action to the assigned id

DoubleOracle.ActionSet — Method

ActionSet(player::Player, actionnames::Vector{String})

Construct ActionSet for player containing action with names as in actionnames.

Examples

julia> ActionSet(Player(1), ["A", "B", "C"])
Actions of Player 1: [1] A | [2] B | [3] C |

julia> ActionSet(Player(2), ["f", "e", "d"])
Actions of Player 2: [1] f | [2] e | [3] d |

Some standard container functions are overloaded to make indexing and iteration simpler.

Base.getindex — Method

getindex(A::ActionSet, id::Integer)

Retrieve name of action with assigned id.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> A[3]
"C"

Base.getindex — Method

getindex(A::ActionSet, name::String)

Retrieve id of action named name.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> A["B"]
2

Base.length — Method

length(A::ActionSet)

Return number of actions in A.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> length(A)
4

Base.iterate — Method

iterate(A::ActionSet)

Go through ids of all actions in fixed order.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> collect(A)
4-element Vector{Any}:
 1
 2
 3
 4

To obtain ::Vector{String} of names of all actions in the set in the exact order as the ids, use the allnames function.

DoubleOracle.allnames — Method

allnames(A::ActionSet)

Return list of action names from A in the same order as ids

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1: [1] A | [2] B | [3] C | [4] D |

julia> allnames(A)
4-element Vector{String}:
 "A"
 "B"
 "C"
 "D"

Another often used feature is selecting a single action at random from the set. This can be done by

DoubleOracle.randomaction — Method

randomaction(A::ActionSet)

Select a random action id of the action set A.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"]);

julia> a = randomaction(A)
a = 3

Game

DoubleOracle.Game — Type

Game

Abstract type uniting all types of games in Game Theory.

First used game is the well-known Normal-Form game represented as an utility matrix. We focus on two-player zero-sum NF games.

DoubleOracle.NormalFormGame — Type

NormalFormGame{T}

Model for a two-player zero-sum game in Normal Form. Here, player 1 tries to maximise the outcome of type T, player 2 minimises it.

Fields

name: Name of the NF game

N: Players of the game

A: Available actions of each player

U: Game matrix containing the outcomes for each joint action profile

There are two possible ways how to create a NormalFormGame structure. First is loading a prepared game from a text file ending with .nfg and following a format given by examples in /data/ folder. Second is setting some parameters and constraints to generate a random game.

DoubleOracle.load — Method

load(filepath::String, type::Type{NormalFormGame})

Load and return a game of type (i.e. $NormalFormGame$, ...) from file located in the filesystem at filepath. The extension of the filepath must correspond to the game type (e.g. $NormalFormGame -> .nfg$).

Example

julia> nfg = load("../data/nf_games/mathing_pennies.nfg", NormalFormGame)

DoubleOracle.generate — Method

generate(::Type{NormalFormGame}, A1min, A1max, A2min, A2max, minutil, maxutil, utilstep)

Generate random Normal-Form game with given parameters.

Fields

A1min::Integer: minimum number of actions of player 1 (default 2)

A1max::Integer: maximum number of actions of player 1 (default 5)

A2min::Integer: minimum number of actions of player 2 (default 2)

A2max::Integer: maximum number of actions of player 2 (default 5)

minutil<:Real: minimum utility possible for player 1

maxutil<:Real: maxium utility possible for player 1

utilstep<:Real: minimum difference between two different utility values for player 1

unique: option to enforce unique payoffs from the interval minutil:maxutil (ignores step to produce sufficient range)

Examples

julia> generate(NormalFormGame; A1min=2, A1max=2, A2min=3, A2max=3, minutil=1, maxutil=1)
===== generated =====
players: Player 1 | Player 2

Actions of Player 1: [1] 1 | [2] 2 |
Actions of Player 2: [1] 1 | [2] 2 | [3] 3 |

U (2 × 3)
    1   2   3
   -------------
 1 | 1 | 1 | 1 |
   -------------
 2 | 1 | 1 | 1 |
   -------------

To avoid always writing nfg.U[a1, a2] a shortcut is provided.

Base.getindex — Method

getindex(nfg::NormalFormGame, a1::Integer, a2::Integer)

Obtain nfg outcome by playing joing action profile (a1, a2)

Examples

julia> nfg = load("../data/nf_games/matching_pennies.nfg", NormalFormGame);

julia> nfg[1, 1]
1.0

julia> nfg["1", "B"]
-1.0