Solution utilities

This part summarizes funtions and structures providing solutions for games defined in previous section.

DoubleOracle.Algorithm — Type

Algorithm

Abstract type representing solving algorithms of a game in the Game Theory

Matrix Game

This serves as a general solver for matrix game with arbitrary size, not necessarily corresponding to size of the actual Normal-Form game. It solves a linear program for the passed matrix u (game or subgame) and saves the outcomes and equilibrium strategies of both players.

DoubleOracle.MatrixGame — Type

MatrixGame

Solve and represent solution of a Matrix game with matrix u. Solution is made by a standard linear program for two-player zero-sum Normal-form games. Both outcome and an equilibrium strategy is saved for both players. Note that this can be only a solution of a subgame, not necessarily the whole NFG.

Fields

outcomes: 2-tuple with game outcomes for each respective player

strategies: 2-tuple with en equilibrium strategies (probability distribution over rows/columns Δ(A)) for each respective player

Examples

julia> u = [1 -1; -1 1];

julia> MatrixGame(u)
MatrixGame results:
→ outcome of the Nash Equilibrium: (0.0, -0.0)
→ strategy of row player: [0.5, 0.5]
→ strategy of column player: [0.5, 0.5]

Again, for easier extraction of results, indexing shortcuts are provided.

DoubleOracle.MatrixGame — Method

(mg::MatrixGame)(p::Player)

Obtain the outcome value of a player from the MatrixGame structure.

Examples

julia> mg = MatrixGame([1 -1; -1 1]);

julia> mg(Player(1))
0.0

Base.getindex — Method

getindex(mg::MatrixGame, p::Player)

Obtain the strategy of a player from the MatrixGame structure.

Examples

julia> mg = MatrixGame([1 -1; -1 1]);

julia> mg[Player(1)]
2-element Vector{Float64}:
 0.5
 0.5

Solution

Solution represents the final solution of the whole Normal-Form game. It holds final equilibria strategies and outcomes.

DoubleOracle.Solution — Type

Solution

Represent solution of the Normal-Form game. In contrast to MatrixGame, this solution corresponds always to the whole NFG.

Fields

name: name of the NormalFormGame (for pretty output)

method: name of the method used to solve the problem

outcomes: 2-tuple of outcome for each player

strategies: 2-tuple of strategies for each player (strategy is a vector of tuples, where first part is name of action and second is a playing probability in equilibrium)

DoubleOracle.Solution — Method

Solution(mg::MatrixGame, nfg::NormalFormGame, method::Type{<:Algorithm})

Construct Solution in a case where mg contains solution of the whole nfg, i.e. in exact solution by linear programming. Strategies in mg are assumed to be of same length as action sets in nfg.

Examples

julia> nfg = load("../data/nf_games/matching_pennies.nfg", NormalFormGame);

julia> mg = MatrixGame(nfg.U);

julia> Solution(mg, nfg, LinearProgram)
The two-player zero-sum Normal-Form game was solved by Linear program
Player 1 gains outcome 0.0 by playing a strategy
 → [1 : 0.5, 2 : 0.5]
Player 2 gains outcome -0.0 by playing a strategy
 → [A : 0.5, B : 0.5]

As always, the shortcuts for easier indexing and uitilization of the Player abstraction.

DoubleOracle.Solution — Method

(solution::Solution)(p::Player)

Shortcut to obtain outcome of a player present in solution.

Examples

julia> solution = Solution("test", LinearProgram, (1.0, -1.0), ([("A", 1)], [("B", 1)]));

julia> solution(Player(2))
-1.0

Base.getindex — Method

getindex(solution::Solution, p::Player)

Shortcut to obtain strategy of a player present in solution.

Examples

julia> solution = Solution("test", LinearProgram, (1.0, -1.0), ([("A", 1)], [("B", 1)]));

julia> solution[Player(2)]
1-element Vector{Tuple{String, Float64}}:
 ("B", 1.0)

When probabilities are to be joined with corresponding action names, the pairstrategies function is advised to be used.

DoubleOracle.pairstrategies — Method

pairstrategies(names::Vector{String}, probs::Vector{Float64})

Merge names and probs (probabilities) in such a way that first element in names is joined with first element in probs into a tuple and so on.

Examples

julia> names = ["A", "B"];

julia> probs = [1.0, 0.0];

julia> pairstrategies(names, probs)
2-element Vector{Tuple{String, Float64}}:
 ("A", 1.0)
 ("B", 0.0)

To compare two solutions whether they are the same (or approximately the same with tolerance atol) use the following functions.

DoubleOracle.samepayoffs — Method

samepayoffs(s1::NTuple{2, Float64}, s2::NTuple{2, Float64}; atol=1e-4)

Compare two solutions s1, s2 whether they have the same (or similar) payoffs. Outcomes are compared with tolerance atol.

Examples

julia> nfg = load("../data/nf_games/matching_pennies.nfg", NormalFormGame);

julia> s1 = solve(nfg, LinearProgram);

julia> s2 = solve(nfg, DoubleOracleAlgorithm);

julia> samepayoffs(s1.outcomes, s2.outcomes)
true

DoubleOracle.samestrategies — Method

samestrategies(s1::Vector{Tuple{String,Float64}}, s2::Vector{Tuple{String,Float64}}; atol=1e-4)

Compare two vectors of strategies s1, s2 whether they are the same including name of actions and order. Probabilities in the strategies are compared with tolerance atol.

Examples

julia> s1 = [("A", 0.2), ("B", 0.5), ("C", 0.3)];

julia> s2 = [("A", 0.20005), ("B", 0.49999), ("C", 0.3)];

julia> s3 = [("A", 0.3), ("B", 0.5), ("D", 0.2)];

julia> samestrategies(s1, s2)
true

julia> samestrategies(s1, s3)
false

DoubleOracle.samesolutions — Method

samesolutions(s1::Solution, s2::Solution; atol=1e-4)

Compare two solutions s1, s2 whether they are the same (or similar). Outcomes, names of actions and probabilities are compared with tolerance atol.

Examples

julia> nfg = load("../data/nf_games/matching_pennies.nfg", NormalFormGame);

julia> s1 = solve(nfg, LinearProgram);

julia> s2 = solve(nfg, DoubleOracleAlgorithm);

julia> samesolutions(s1, s2)
true

For further analysis of the solution, the functions related to strategy support can be used.

DoubleOracle.support — Method

support(strategy::Vector{Tuple{String,Float64}})

Compute support of a strategy, i.e. set of actions played with non-zero probability in the strategy. Return the list of names of the actions.

Example

julia> strategy = [("A", 0.1), ("B", 0.5), ("C", 0.0), ("D", 0.0), ("E", 0.4)];

julia> support(strategy)
3-element Vector{String}:
 "A"
 "B"
 "E"

DoubleOracle.support — Method

support(s::Solution)

Compute support of a all strategies in the Solution s. Returns an NTuple{Vector{String}} of action names based on the number of strategies N in s.

Example

julia> nfg = load("../data/nf_games/test.nfg", NormalFormGame);

julia> s = solve(nfg, LinearProgram);

julia> support(s)
(["1", "2"], ["B", "C"])

DoubleOracle.supportratio — Method

supportratio(strategy::Vector{Tuple{String,Float64}}; digits=2)

Compute ratio of actions in the support of a strategy to the total number of available actions. Return number from range $[0, 1]$. The digits keywoard parameter specifies the rounding precision of the ratio.

Example

julia> strategy = [("A", 0.1), ("B", 0.5), ("C", 0.0), ("D", 0.0), ("E", 0.4)];

julia> supportratio(strategy)
0.6

DoubleOracle.supportratio — Method

supportratio(s::Solution; digits=2)

Compute the support ratio for all strategies in the Solution s. Returns an NTuple{Float64} based on the number of strategies N in s. The digits keywoard parameter specifies the rounding precision of the ratio.

Examples

julia> nfg = load("../data/nf_games/test.nfg", NormalFormGame);

julia> s = solve(nfg, LinearProgram);

julia> supportratio(s)
(1.0, 0.67)