Solution by Double Oracle algorithm

This section discusses solution of a two-player zero-sum Normal-Form game by using the Double Oracle algorithm.

DoubleOracle.DoubleOracleAlgorithm — Type

DoubleOracleAlgorithm

Subtype of Algorithm representing solution of a Normal-Form game by the Double Oracle algorithm.

First, we describe some necessary structures and functions which enable us solving the game.

Oracle

First and most important concept for this solution mechanism is an Oracle. It tells which actions are likely to be a support of the equilibrium mixed strategy.

DoubleOracle.Oracle — Type

Oracle

Keep strategies already present in the restricted Normal-Form game. Changes occur between iterations and thus represents (partial) state of the Double Oracle algorithm.

Fields

player: owner of the oracle

selected: vector of action ids of a given player which are present in the restricted game

Base.length — Method

Base.length(O::Oracle)

Return number of actions in the oracle O.

Examples

julia> O = Oracle(Player(2), 3)
Oracle for Player 2 has actions: [3]

julia> length(O)
1

Base.iterate — Method

Base.iterate(O::Oracle)

Provide a way to iterate over selected actions in the Oracle O.

Examples

julia> O = Oracle(Player(2), 1);

julia> add!(O, 2);

julia> add!(O, 4);

julia> collect(O)
3-element Vector{Any}:
 1
 2
 4

To add another action id to the Oracle, the prepared modifier function can be used.

DoubleOracle.add! — Method

add!(O::Oracle, actionid::Integer)

Add a new actionid of an action into the selected actions in Oracle O.

Examples

julia> O = Oracle(Player(2), 1)
Oracle for Player 2 has actions: [1]

julia> add!(O, 2);

julia> O
Oracle for Player 2 has actions: [1, 2]

Game restrictions

The Double Oracle algorithm always operates over some subgame to reduce computation time. These subgames are defined by one or both oracles and are different in each situation.

DoubleOracle.GameRestriction — Type

GameRestrictions

Abstract type uniting possible partial restrictions of the game. Partial restriction means only some subset of rows or columns is preserved.

DoubleOracle.ColumnRestriction — Type

ColumnRestriction

A game restriction which preserves only some columns from the payoff matrix. All rows in the given columns are preserved as well.

DoubleOracle.RowRestriction — Type

RowRestriction

A game restriction which preserves only some rows from the payoff matrix. All columns in the given rows are preserved as well.

These restrictions are then created from the full payoff matrix of the original nfg game and created Oracles. The type of restriction is given by the above mentioned restriction types.

DoubleOracle.restrict — Method

restrict(nfg::NormalFormGame, O1::Oracle, O2::Oracle)

Create a restriction of the payoff matrix of nfg, where only a subset of columns and a subset of rows is preserved. The preserved rows are given by actions in Oracle O1, the columns by Oracle O2.

Examples

julia> nfg = load("../data/nf_games/test.nfg", NormalFormGame);

julia> O1 = Oracle(Player(1), 1);

julia> O2 = Oracle(Player(2), 2);

julia> add!(O2, 3);

julia> nfg.U
2×3 Matrix{Float64}:
  30.0  -10.0   20.0
 -10.0   20.0  -20.0

julia> O1
Oracle for Player 1 has actions: [1]

julia> O2
Oracle for Player 2 has actions: [2, 3]

julia> restrict(nfg, O1, O2)
1×2 Matrix{Float64}:
 -10.0  20.0

DoubleOracle.restrict — Method

restrict(nfg::NormalFormGame, O::Oracle, ::Type{ColumnRestriction})

Create a restriction of the payoff matrix of nfg, where only a subset of columns is preserved. The preserved columns correspond to action ids given by Oracle O.

Examples

julia> nfg = load("../data/nf_games/test.nfg", NormalFormGame);

julia> O = Oracle(Player(2), 2);

julia> add!(O, 3);

julia> nfg.U
2×3 Matrix{Float64}:
  30.0  -10.0   20.0
 -10.0   20.0  -20.0

julia> O
Oracle for Player 2 has actions: [2, 3]

julia> restrict(nfg, O, ColumnRestriction)
2×2 Matrix{Float64}:
 -10.0   20.0
  20.0  -20.0

DoubleOracle.restrict — Method

restrict(nfg::NormalFormGame, O::Oracle, ::Type{RowRestriction})

Create a restriction of the payoff matrix of nfg, where only a subset of columns is preserved. The preserved columns correspond to action ids given by Oracle O.

Examples

julia> nfg = load("../data/nf_games/test.nfg", NormalFormGame);

julia> O = Oracle(Player(1), 1);

julia> nfg.U
2×3 Matrix{Float64}:
  30.0  -10.0   20.0
 -10.0   20.0  -20.0

julia> O
Oracle for Player 1 has actions: [1]

julia> restrict(nfg, O, RowRestriction)
1×3 Matrix{Float64}:
 30.0  -10.0  20.0

Best response

Last necessary component of the Double Oracle algorithm is a concept of a best response. Here, the players search for an optimal pure strategy (i.e. single action), which would result in the best payoff (from their perspective) while keeping the Oracle and strategy of the opponent fixed.

DoubleOracle.bestresponse — Method

bestresponse(nfg::NormalFormGame, O::Oracle, p::Vector{Float64}, ::Type{ColumnRestriction})

Find id of a best response pure strategy in a Normal-Form game, when fixing Oracle O of the opponent and the equilibrium mixed strategy p, which is computed by solving a game restricted by Oracles of both players. This function finds the best response of the row player.

DoubleOracle.bestresponse — Method

bestresponse(nfg::NormalFormGame, O::Oracle, p::Vector{Float64}, ::Type{RowRestriction})

Find id of a best response pure strategy in a Normal-Form game, when fixing Oracle O of the opponent and the equilibrium mixed strategy p, which is computed by solving a game restricted by Oracles of both players. This function finds the best response of the column player.

Now, all the components for the complete Double Oracle algorithm are prepared.

Double Oracle algorithm

This algorithm is more suitable for games with large number of actions, where it is not expected that the support of an equilibrium strategy is not expected to be very large. On such games, it should outperform the standard Linear programming method.

DoubleOracle.solve — Method

solve(nfg::NormalFormGame, ::Type{DoubleOracleAlgorithm})

Solve the NormalFormGame by Oracle algorithm and return the payoff and equilibrium strategies.

Algorithm

Start from random actions in the Oracles for both players

Repeat

Solve a restriction of the game given by the Oracles by Linear programming method

Find best response pure strategies to the equilibrium strategies from the restricted game

If both found pure strategies are already present in the Oracles, end the loop

Add best responses to Oracles

Return final equilibrium strategies and payoffs

Note that the double oracle algorithm does not necessarily compute the whole strategy, but only a part corresponding to the restricted payoff matrix. To fix this issue, the following function is used to finalize the output.

DoubleOracle.fullstrategy — Method

fullstrategy(O::Oracle, p::Vector{Float64}, n::Integer)

Creates a new strategy of length n. The probabilities of actions in the Oracle O are replaced with values from restricted strategy p (of the length n). Rest of the probabilities are set to 0.0.

Examples

julia> O = Oracle(Player(1), 1);

julia> add!(O, 3);

julia> O
Oracle for Player 1 has actions: [1, 3]

julia> p = [0.3, 0.7];

julia> fullstrategy(O, p, 4)
4-element Vector{Float64}:
 0.3
 0.0
 0.7
 0.0