Exact solution by linear programming

DoubleOracle.LinearProgram — Type

LinearProgram

Subtype of Algorithm representing exact solution of a Normal-Form game by linear programming.

To obtain an exact solution of a given Normal-Form game, use the function solve with LinearProgram parameter.

DoubleOracle.solve — Method

solve(nfg::NormalFormGame, ::Type{LinearProgram})

Solve the NormalFormGame by linear programming and return the outcome and equilibrium strategies.