DoubleOracle

Documentation for DoubleOracle.

DoubleOracle.ActionSet

DoubleOracle.ActionSet

DoubleOracle.Player

Base.getindex

Base.getindex

Base.iterate

Base.length

DoubleOracle.createplayers

Documentation

DoubleOracle.ActionSet — Type

ActionSet

Set of actions of a single player. Stores names and assigns ids for internal representation.

DoubleOracle.ActionSet — Method

ActionSet(player, actionnames)

Construct ActionSet for player containing action with names as in actionnames.

Examples

julia> ActionSet(Player(1), ["A", "B", "C"])
Actions of Player 1
→ [1] A
→ [2] B
→ [3] C

julia> ActionSet(Player(2), ["f", "e", "d"])
Actions of Player 2
→ [1] f
→ [2] e
→ [3] d

DoubleOracle.Player — Type

Player

Represent a Player with an integer id in a Game

Examples

julia> Player(1)
Player 1

julia> Player(2)
Player 2

Base.getindex — Method

getindex(A, id::Integer)

Retrieve name of action with assigned id.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1
→ [1] A
→ [2] B
→ [3] C
→ [4] D

julia> A[3]
"C"

Base.getindex — Method

getindex(A, name::String)

Retrieve id of action named name.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1
→ [1] A
→ [2] B
→ [3] C
→ [4] D

julia> A["B"]
2

Base.iterate — Function

iterate(A)

Go through ids of all actions in fixed order.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1
→ [1] A
→ [2] B
→ [3] C
→ [4] D

julia> collect(A)
4-element Vector{Any}:
 1
 2
 3
 4

Base.length — Method

length(A)

Return number of actions in A.

Examples

julia> A = ActionSet(Player(1), ["A", "B", "C", "D"])
Actions of Player 1
→ [1] A
→ [2] B
→ [3] C
→ [4] D

julia> length(A)
4

DoubleOracle.createplayers — Method

createplayers(n)

Create a vector of $n$ Players with ids $i ∈ {1,…,n}$

Examples

julia> createplayers(3)
3-element Vector{Player}:
 Player 1
 Player 2
 Player 3